requirements.txt
asgiref>=3.7.2,<4.0.0
cffi>=1.15.1,<2.0.0
cryptography>=41.0.3,<42.0.0
Django>=3.2.4,<4.0.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
Pillow>=9.5.0
psycopg2-binary>=2.9.7,<3.0.0
pycparser>=2.21,<3.0.0
PyJWT>=2.8.0,<3.0.0
pytz>=2023.3,<2024.0.0
sqlparse>=0.4.4,<0.5.0
typing_extensions>=4.7.1,<5.0.0
gunicorn
django-cors-headers
dj_database_url